[ADD] #253 department filters to multiple models
