+++ ps_planning/models/__init__.py
@@ -1,6 +1,7 @@
 # Copyright 2024 Hunki Enterprises BV
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl-3.0)
 
+from . import ps_planning_department_mixin
 from . import project_project
 from . import ps_planning_line
 from . import ps_contracted_line
